move functions dir
