+++ src/context.ts
@@ -1,7 +1,6 @@
 import { ArgumentParser, Assign, Execute, GetName } from "@/@types/execute";
 import { CommentMapper } from "@/commentMapper";
 
-let execute: Execute;
 let argumentParser: ArgumentParser;
 let assign: Assign;
 let getName: GetName;
@@ -12,10 +11,6 @@ let currentTime = 0;
 
 let comments: CommentMapper[] = [];
 let isWide = false;
-
-const setExecute = (val: Execute) => {
-  execute = val;
-};
 
 const setArgumentParser = (val: ArgumentParser) => {
   argumentParser = val;
@@ -44,7 +39,6 @@ const setComments = (val: CommentMapper[]) => {
 const setIsWide = (val: boolean) => (isWide = val);
 
 export {
-  execute,
   argumentParser,
   assign,
   getName,
@@ -52,7 +46,6 @@ export {
   currentTime,
   comments,
   isWide,
-  setExecute,
   setArgumentParser,
   setAssign,
   setGetName,
