+++ src/core/core.ts
@@ -1,5 +1,6 @@
 import { A_ANY, T_scope } from "@/@types/ast";
 import { IFunction, IFunctions, INiwangoCore } from "@/@types/core/core";
+import { execute } from "@/core/coreContext";
 
 class NiwangoCore implements INiwangoCore {
   private functions: IFunctions;
@@ -11,5 +12,9 @@ class NiwangoCore implements INiwangoCore {
     this.functions[func.name] = func;
   }
 
-  run(niwango: A_ANY, scopes: T_scope[]) {}
+  run(niwango: A_ANY, scopes: T_scope[]) {
+    execute(niwango, scopes);
+  }
 }
+
+export { NiwangoCore };
