+++ src/core/coreContext.ts
@@ -1,0 +1,16 @@
+import { Execute } from "@/@types/execute";
+
+let execute: Execute;
+
+const setExecute = (val: Execute) => {
+  execute = val;
+};
+
+const prototypeScope = {
+  Array: {},
+  String: {},
+  Object: {},
+  Number: {},
+};
+
+export { execute, setExecute, prototypeScope };
