+++ src/core/functions/distance.ts
