+++ src/core/functions/drawText.ts
