+++ src/core/functions/index.ts
@@ -2,7 +2,7 @@ import { processDump } from "./dump";
 import { processDef } from "./def";
 import { processDefKari } from "./def_kari";
 import { processWhileKari } from "./while_kari";
-import { A_CallExpression } from "@/@types/ast";
+import { A_CallExpression, T_scope } from "@/@types/ast";
 import { processTimes, timesCondition } from "./times";
 import { processDrawText } from "./drawText";
 import { processDrawShape } from "./drawShape";
@@ -14,13 +14,19 @@ import { processDistance } from "./distance";
 import { processScreenHeight, processScreenWidth } from "./screen";
 import { processTimethis } from "./timethis";
 
+export type IrFunction = (
+  script: A_CallExpression,
+  scopes: T_scope[],
+  object: { [key: string]: unknown }
+) => unknown;
+
 type IFunctions = {
   [key: string]:
     | {
-        func: Function;
+        func: IrFunction;
         condition: (object: unknown, script: A_CallExpression) => boolean;
       }
-    | Function;
+    | IrFunction;
 };
 
 const functions: IFunctions = {
