+++ src/core/functions/times.ts
@@ -1,5 +1,5 @@
 import { A_ANY, A_CallExpression, Argument, T_scope } from "@/@types/ast";
-import { execute } from "@/context";
+import { execute } from "@/core/coreContext";
 
 /**
  * @関数
