+++ src/core/functions/timethis.ts
@@ -1,5 +1,5 @@
 import { A_CallExpression, T_scope } from "@/@types/ast";
-import { execute } from "@/context";
+import { execute } from "@/core/coreContext";
 
 const processTimethis = (script: A_CallExpression, scopes: T_scope[]) => {
   console.time("timethis");
