+++ src/core/processors/BinaryExpression.ts
@@ -1,5 +1,5 @@
 import { A_BinaryExpression, T_scope } from "@/@types/ast";
-import { execute } from "@/context";
+import { execute } from "@/core/coreContext";
 import { NotImplementedError } from "@/errors/NotImplementedError";
 import {
   Addition,
