--- src/core/processors/CallExpression/index.ts
@@ -1,1 +1,0 @@
-export * from "./CallExpression";
