+++ src/core/processors/LogicalExpression.ts
@@ -1,5 +1,5 @@
 import { A_LogicalExpression, T_scope } from "@/@types/ast";
-import { execute } from "@/context";
+import { execute } from "@/core/coreContext";
 import { NotImplementedError } from "@/errors/NotImplementedError";
 
 /**
