+++ src/core/processors/UnaryExpression.ts
@@ -1,5 +1,5 @@
 import { A_UnaryExpression, T_scope } from "@/@types/ast";
-import { execute } from "@/context";
+import { execute } from "@/core/coreContext";
 import { NotImplementedError } from "@/errors/NotImplementedError";
 import { BitwiseNOT, LogicalNot, UnaryNegation, UnaryPlus } from "@/operators";
